FEATURE/SORTOF/NOTREALLY/SHOOTMENOW: ?search=term will do a search on dataset page.
QA: This is a terrible feature that hopefully no one will use.
The user experience on it will suck. But, if you add ?search=your%20search%20term
to the end of a dataset grid page, it will automatically do a search when the page
is first loaded. The first time you clear search, the entire page will reload and
the search should be gone. You should be able to clear the temp view as well, but
if you refresh the page at any point before explicitly clicking clear search, the
search will come back. It's awful, but someone apparently needs it.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>