BUG: Requests were not setting domain properly
The core server has been getting gobs of requests for the URL
/domains/.json, indicating that the frontend is trying to retrieve an
empty current domain configuration. This could in theory happen with
HTTP/1.0 requests, but nobody ever uses HTTP/1.0 anymore. (That's like
Netscape Navigator 1.0 old)

My best guess as to a fix is that the proxy is sending an
HTTP_X_FORWARDED_HOST header, but it's empty. I'm being a bit more
careful in selecting the host header now, and logging a bit more info
when things are broken. I also fixed a latent bug where I wasn't
stripping the port from the hostname.

QA: Not really sure how to test it. If you notice anything weird with
the a domain not being set properly (theme leakage and the like) let me
know.

Reviewed-by: Robert Macomber <robert.macomber@socrata.com>