BUG: Login cookie fixes
This fixes two separate things:
1. We were incorrectly determining when to set the remember_token cookie
that allows you to close your browser and remain logged in.
2. Once that "remember me" functionality worked again, I fixed a bug
where the user would get logged in but the logged_in cookie that toggles
the links in the upper right via Javascript wasn't being set with it.

QA: "Remember me" should work again - specifically, you can login with
it checked, close your browser completely, and start it up again and be
logged back in to the site with the same account. Also, check that the
"Sign in/Sign out" links in the top-right reflect your signed-in status.

Reviewed-by: Chi Tang <chi.tang@socrata.com>