BUG: Sometimes you appear logged in when you're really not.
The repro steps for this is unfortunately really boring:
1. Sign in
2. Leave your browser open, but don't use the site for 6 hours
3. Visit a page requiring authentication. Notice that you get logged
out, but you still see the header in the top-right with the logged-in
state instead of the logged-out state.

The issue was that the core server session credentials have a signed
timestamp after which they're no longer valid, regardless of the
cookie's expiration timestamp. (This is useful to limit replay attacks,
among other things.) However, even though the cookie was no longer
valid, we weren't removing the logged_in cookie properly. This change
adds a before_filter to manage that.

QA: See repro steps above. It might be useful to use a different browser
than your normal one, or use a VM you suspend for most of the day with
the browser remaining open so you can do other things in the meantime,
since you have to wait 6 hours for the cookie to become invalid. You
cannot cheat this by changing your system clock; the time is validated
server-side rather than client-side.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>