SPARTA!: MURDERING OLD WIDGETS.
The old SDP no longer exists whatsoever. Any widgets, regardless
of whether they were /black or /whitehouse or /4-4 of old theme,
will now render as the new widget.

The new widget will politely ignore anything you throw at it
that is malformed, and it will just render the default theme.
This includes old widget default, old widget 4-4, /black,
/whitehouse, /whatever.

Note that there's some cruft like the meta tab partials that
could be drastically simplified as a result of this, but given
that we're going to kill the old datasets page in a matter of
weeks, it doesn't seem worth it to deal with untangling it all.

QA: Anything that previously rendered as the old widget will
now render with the new widget. Unless a theme was explicitly
created for SDP refresh, it will simply fall back to the
default.

Needless to say, this requires a full regression on everything
to do with SDP's.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>