BUG 1895: RDF list/custom button not clickable.  Set cursor of toggle button to pointer.
QA: Combobox in IE8 has a width equal to the current option. In FF or IE7, it is the widest option. This causes the problem.
Also remove the unnessary (confusing to me) usage of "self" in view.rb.  Self sometimes refer to the class while at other times it refers to the current object.

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>