BUG 2019: Sometimes, federated dataset url is pointing to the current domain.
QA: URL always point to the right domain.
We think this is a cache problem that we fixed.  But if you do see this again,
please view source or find element in Firebug or IE debugger.  We add a debugging attribute to the <a> tag named "_domain" in staging.  Please attach this debugging info along with bug updates.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>