SPARTA: Upgrade Hoptoad notifier to 2.3.5
After upgrading to Rails 2.3.8, a large number of our deprecation
warnings come from Hoptoad. We needed to upgrade anyways (so we stop
getting nagged in the notification emails) so now seemed like a good
time to do so.

Hoptoad did change from being a plugin to being a gem. To make this work
with bundler, I deleted the vendor/plugins/hoptoad_notifier, added the
gem to our Gemfile, and then ran:

  bundle install
  bundle exec script/generate hoptoad --api-key $API_KEY

The script/generate failed initially because of the way we're assuming
that a @@current_domain is set; the changes to application_controller
introduces a fallback for rendering an error when the error is related
to finding the domain for the request in the first place.

QA: We should still get hoptoad errors as before; remember the
/test_page/kaboom URL I made way back when to force an error when you
can't find any other way to cause an exception.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>