BUG: Unable to POST JSON payloads in Ruby 1.9.2
Ruby 1.9.2 finally got smart about Unicode and strings vs. byte sizes,
and the new net/http code chokes when we pass it a JSON object as the
body of a Net:HTTP request. It's now necessary to explicitly convert the
payload into a string before setting it as the body.

QA: I reproduced this by trying to add a module to a domain in /internal
- you'll still be testing on Ruby 1.8.7 though, so it's mostly a matter
  of ensuring it didn't regress.

Reviewed-by: Robert Macomber <robert.macomber@socrata.com>