BUG: Keep wizard tips hidden when dialogs are shown * Ref count open dialogs in sidebar
QA:
As an unsigned in user, create a temp view, make sure the sidebar is
open with a wizard tip shown.  Choose Save As... -- the wizard tip
should hide. Give the view a name; this should show you a sign-in
dialog.  Hit 'No account yet' to get the sign-up dialog -- the wizard
should still be hidden; hit Cancel; make sure the wizard tip doesn't
reappear until all the dialogs are closed.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>