BUG: Inline login/signup breakage on the new dataset page was broken
Our code for capturing navigation away from the page was storing off the
AJAX form submission url to /signup and redirecting there after creating
the account.

In the future, if you have a link that you don't want to be subject to
the saved-redirect feature, add the class 'noRedirPrompt' to the a
element.

QA: Logout. Visit a public dataset, and create a filter. Try to save the
filter. Before, you'd notice a few weird issues - "Forgot password"
would pop up a dialog box under the login box you couldn't read.
"No account yet" would create an account and log you in, but redirect
you to /signup instead of the newly-saved dataset URL.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>