BUILD: Bundler cleanup to canonical usage
According to the bundler documentation at
http://gembundler.com/deploying.html we're not supposed to have anything
in the .bundle directory under source control. With the bundler upgrade
to 1.0, we're going to change the bundle location to vendor/bundle
instead of vendor/bundler_gems, so now's a good time to kill off that
old .bundle/config file.

QA: Nothing to test.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>