BUILD: Auto-setup RVM to use 1.9.2 and a custom gemset
For those of us using RVM (which should be most of us at this point),
this change makes it so that you can just cd into the frontend source
code and have RVM automagically switch you to use Ruby 1.9.2 and a
custom gemset (thereby removing any conflicts with system gems)

This, combined with bundler, sets us up so that we're always ensuring
we're working with as clean an environment in development as possible.

If you don't have RVM installed/configured, nothing will happen.

QA: Nothing to test; this only affects development

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>