FEATURE: UI to indicate when a non-matching row is created * Use the _underlying flag from the core server to mark a non-matching row * Mark the row as grey, X for row number
QA:
Open a filter, create a row that does not match it.  The row should
turn grey, and instead of a row number it should have an 'X' with a
tooltip.  Modify the row to match the filter (then make one more edit
after that due to a core server bug); the row should now be white and
have a row number

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>