BUG: Frontpage could possibly cache the wrong state.
Basically, I always knew it was possible that the homepage would cache
the wrong state for the browse control, but figured it wasn't a case
worth covering given that the only way to do it would be to deliberately
much with the queryparams manually. It turns out that search engines and
some websites still remember and link to the *old* homepage urls, and
the queryparams from old to new discover are *just* similar enough that
it'll affect the rendering of the browse control. I add an option that's
activated on the homepage only to the browse param processing to force
it to use the default params. Essentially, this means that if you do
attempt to give queryparams to the homepage, it'll just ignore them and
hand you the default homepage.

This is potentially bad for SEO, but since there should nothing still
linking at those old urls of significance, I figure they'll soon give
those hits up as dead anyway and it'll cease to be a problem after a day
or two.

QA: If you feel really dedicated, coordinate between the two of you and
make sure you don't hit the homepage of some domain for at least ten
minutes; ideally fifteen to twenty minutes just to be safe. Then, hit it
with ?page=2 on the URL (eg for demo.test:
http://demo.test.socrata.com/?page=2). Make sure it renders page 1
anyway. Then, remove the ?page=2 and go to the homepage normally, and
make sure it also still renders page 1.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>