FEATURE: Real rendering of data in fat row * Moved all rendering/layout code into richRenderer * Rewrote layout code to handle arbitrary number of columns * Tweak fatrow styling
QA:
There should now be a button between 'View as table' and 'View as
single row' called 'View as rich list' which gives a three-column view
kind of like page view, but with 20 rows at a time in a scrollable
list.  There is then paging at the bottom to go through the rows a
chunk at a time.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>