FEATURE: Stories customizer.
QA: Under Homepage, there is Stories Appearance =>
Customize Now. The interface should look familiar,
and it should be pretty self-explanatory. The only
thing to look out for is to make sure you also try
it on a domain that has no stories yet; you should
then get an interface with some sample stories and
you should still be able to customize it as usual.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>