FEATURE: Enable user search on /admin/users page by email. BUG: Don't display current search query as title text.
QA: Remind me to reindex users, because I'll probably have
forgotten to. Then, try to search for a user on the
admin/users page by email and verify that results show up.

Also, your search term should no longer appear grey in the
search box and disappear when you click on it.

Reviewed-by: Robert Macomber <robert.macomber@socrata.com>