BUG: Fix wrapping in page view If a column title was too long, it would wrap to two lines.  Then, if the data was too long and also wrapped, the second line would wrap fully below the title, leaving a long gap between the first and second lines of data.  Changing inline-block to float:left makes that wrap correctly; and doesn't seem to affect anything else in the layout
QA:
Look at page view in a dataset with a long column title and long
column data (just text).  The column title should wrap to two lines,
and the data should also wrap to multiple lines.  Make sure the data
wraps properly, so that there isn't a large vertical gap between the
first and second lines of the data.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>