FEATURE: Allow the ability to dismiss the maintenance message by clicking a close button
There are two noticeable changes here:
1. The downtime message now shows up for 15 seconds instead of 5 on
every page.
2. The downtime message now has a close box you can click to dismiss it.
It saves a session cookie, which means that you won't see the downtime
message again until the downtime message changes (we redeploy with a new
time window or message) or you quit your browser entirely and restart.

As an added plus, the HTML for the downtime is now injected in the
Javascript, which means that it's much less likely to get picked up by
search engines as part of the site's snippet text.

QA: You'll need to have the downtime window set up on test for this. It
currently expires this Saturday, so we'll have to figure out a day to
manually override and test this.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>