FEATURE: Link directly from map create to location column create if no location already * Instead of disabling map pane, show an error message * Link to column create, pre-set with location type and possibly one or both 'Use existing...' sections expanded * After adding a column that enables map create, show the map create pane * Add a once listener option to base-model.js * If there is a hidden loc column, then link to show/hide
QA:
In a dataset without a location column, the map create pane should now
show an error message and provide links that open the Add Column pane
with Location already selected.  The different links should also
expand 'Use existing...' sections as appropriate.  After creating a
location column, the Map create pane should re-open and be enabled.
In a view or dataset where you can't create a column, only a short error
message without links is displayed. In a view with a hidden location
column, it should provide a link to the show-hide pane.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>