FEATURE: Guess at defaults for column selects in sidebar * Add an option for column selects that takes a list of names to use to guess the default column * Use this for location convert, calendar create, map create * For all column selects, if there is only one column, then choose that by default
QA:
For some column selects in location create, map create and calendar
create, it will now guess at defaults by looking at column names. For
example, when converting columns into a location column, the Latitude
option will look for a column named 'Latitude', 'Lat', or 'Y' and set
that as the default if it found one. All the selects in location convert
do this; all the column selects in calendar create do this; only the
title & description selects in map create do this.

Separately, for all column selects, if there is only one column
available, it  will select that column by default.

All of this is just to make it a bit easier for users to go through the
dialogs. They can always select a different option, and we're not always
going to guess the right column.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>