SECURITY: Don't allow browsers to autocomplete password fields
This is debatable whether it's really a security vulnerability - if a
client cares about the security of their system, they have all the tools
available to make their browser not remember password fields. In
general, I'm supportive of password managers and don't want to prohibit
their usage, but this particular case is for older browsers (IE6, I'm
looking at you) where you start filling in a form and it autocompletes
the entire form for you - including password fields. Modern versions of
IE (and older versions of IE using SSL pages) don't do this, nor do any
other modern browser, so this is mostly bogus.

Still, it shouldn't hurt to do this and it's a small, localized change
that affects all password fields easily.

QA: Forms with password fields should continue to work, but you may
notice that the browser no longer autocompletes the password field for
you, if it did before. Note that it WILL still autocomplete if you use a
password manager, such as the one built in to Firefox or Chrome, or
3rd-party add-ons to IE.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>