FEATURE: Unified filter should be cleverer about when it does/doesn't pop the unsaved view message.
QA: Make sure that the unified filter doesn't pop
the unsaved view message when you do something that
isn't really an actual operation (eg change the column
of a condition that has zero values to filter on, or
click into and out of the text field of a text column
filter condition). More importantly, make sure it DOES
pop when you DO do something that should change the filter.

To-be-
Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>