BUG: Make it so that the grid on a map ignores the viewport boundaries * Named filters can now take an array of displayTypes they apply to. If looking at a non-matching displayType, then skip that named filter * Detect when the displayType changes and a named filter becomes ir/relevant, and trigger a query_change * Removed a clause where named filters wouldn't get copied back over to the view if the server query came back empty. This was blowing away named filters for non-matching displayTypes. I'm not sure exactly why this was added in the first place; I'm hope it isn't important...
QA:
Load up a map with more than 500 points. Zoom until you have only a
subset viewable (the red warning is gone). Switch to the grid, and
make sure it shows all the rows. Switch back to the map, and make sure
it has no warning and shows you points.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>