BUG: Suggested datasets button wrapping poorly in IE7.
QA: In the "Didn't find what you're looking for?" box
at the bottom of the homepage and browse page, the Suggest
button used to appear terribly broken IE7. Please verify
that it looks okay in all browsers now. Same with the
Create a New Dataset button in Profile. I couldn't find
a ticket for the bug; could be filed under Jeff?

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>