FEATURE: When you first hit the unified filter, display some friendly text.
QA: When you hit the unified filter, there should be
a message offering to send you to a tutorial video.
The video page doesn't exist yet, so it's expected
that it'll throw a 404 not found when you try to view
it. Make sure that
1. When you click on the add button, that a condition
is added and the filter starts functioning as before.
2. When there is already a default or saved filter, the
friendly message doesn't show up; the filter should 
behave as before.
3. When you click on the edit default filter button
with the friendly message visible, it should go away.

Reviewed-by: Chris's Time Machine <chris.metcalf@socrata.com>