FEATURE: Video gallery in popup as well as at /videos
QA: If you go back to that video link in the filter
sidebar, when you click on it now it should instead
open up in a popup window, and show you "related"
videos (what it's doing is filtering by the same
category and audience, and sorting on "basics first"
sort, and taking the 3 videos around the one you're
watching, but since we only have 3 videos it doesn't
really do anything useful yet). You should be able
to click on see more to get to /videos, and going
directly to /videos should still look as it did before.

To-be-
Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>