FEATURE: CSRF protection in Rails+JS.
As of Rails 2.3.11 (and in 3.0.4), the CSRF protection behavior
has changed. No longer will it balk and refuse to service your
request -- instead, they'll clear your user session and continue
on their merry, and so the request will be serviced as if you had
been logged out. This would be fantastic! if we used Rails
sessions to keep track of user auth. But we don't. So we have to
monkey patch in some shenanigans.

In addition, in Rails 2.3.11, xhr requests are no longer
automatically whitelisted by the CSRF protection checker. This
means that all JS requests must also include valid CSRF credentials.
This was done by dumping the CSRF authenticity token as a meta tag
and including it as a X-CSRF-Token HTTP header on every AJAX request
back to either Rails or the Core Server. It also includes it on GETs,
which is strictly unnecessary, but that's okay -- it'll just get duly
ignored.

The core server does not yet feature protection on this stuff, but
there is also a bit of accounting for that -- when the user loads a
page on the website, it'll automatically take the CSRF token from the
meta tag and dump it into a 'socrata-csrf-token' cookie. We have to
do this because even though the Rails session (and thus the csrf
token) is stored in the cookie, it's marshalled and encrypted so that
the Core server can't really read it. In the future, the Core server
will require that requests on non-https that are attempting to auth
via a cookie will *require* a header of X-CSRF-Token as well as a
cookie of 'socrata-csrf-token' that match. Between these and the fact
that the user auth token is being submitted through a cookie as well,
it should be enough to verify that the user is legitimately requesting
through our app.

QA: Regression. What would break would be Javascript requests back
to Rails, which is in practice a very small number of requests. If a
request fails, and you get mysteriously logged out afterwards, then
something has gone wrong to do with this change.

If you're feeling ambitious:
Go to a page like edit_metadata on a dataset. Use Firebug; inspect one
of the form fields (right-click on it; Inspect Element). You'll get a
bunch of HTML that comes up. You want to find the one that looks like:

<div style="margin: 0pt; padding: 0pt; display: inline;">

It should be nested *just* inside the <form> tag. Expand it, and there
will be an <input> tag inside of that, and it should have something like

value="83eCm4fCu7oUOytwhxvgK6OV3t6UvzhfPFA28x03ZpA="

Double-click on that random string and change it to whatever you'd like;
doesn't matter what as long as you change it (or blank it). Submit the
form. You should get an error message and be logged out.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>