SECURITY 3499: Cross-Frame Security (frame-busting).
QA: Visit /test_page/iframe and /test_page/frameset on
any Socrata test domain. If after some thinking the browser
does not forward you back to /, something has gone wrong
with this change. So for instance,

http://opendata.test-socrata.com/test_page/iframe

should forward back to

http://opendata.test-socrata.com

Please verify that this works in all browsers. Also check
to make sure there aren't any pages on the site that are
caught in an endless refresh loop.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>