FEATURE: Rich layout editor: Ability to toggle relative/absolute sizes * Added control menu with options to set width as relative (%) or absolute (em) * Menu has options to reset height & width to auto/default * Show appropriate size in resize box * Moved item remove into control menu
QA:
Rich layout editor:
Removing an item should now happen from a menu on each item, not a button you drag to.
Items that can be resized should now have options in the menu to
toggle between relative and absolute widths. You should also be able
to reset the height & width to default. It should show the appropriate
size in the gray box during resize (numbers in percent, like 50% when
size is relative; in px when absolute).

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>