FEATURE: Routing & Approval queue * Basic structure for routing & approval under admin * Render list of datasets that need approval * Paging & search * Expanded info with approval history * Approve & Reject actions that work * Approve model in JS, approval extension to Dataset * Add method to load a user to User
QA:
Once routing & approval has been enabled, you should get a section in
your admin section.  The Approval Queue section should work.  It will
list datasets, and should show the current status of approval, and who
is next up to approve.  It should have actions for Approve or Reject,
as appropriate for what stages your user is in for approving.  See the
spec for more details.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>