FEATURE: daterangepicker wouldn't autoaccept current view when hitting accept.
QA: On admin/analytics page, click on the "Mar 01, 2011 - Mar 21, 2011"
range selector at the top, and choose either Month Of... or Year Of...
Change the range to a different month or year via the arrows or the
dropdown, and press Apply. Make sure the date range updates to what
you were looking at.

Previously, you'd have to click on a day within the month in order
to make it accept it. That's pretty awkward.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>