BUG: Perf improvements for rich list, especially IE In a dataset with lots of columns, rich list was really slow no matter how many columns you had configured, because it tried to fill in data for all the columns. Change it to only use the columns that are actually rendered.
QA:
In a very wide dataset (over 100 columns), configure rich list to only
show 10 columns are so, and save that as the default view. Load it up
in IE, and it should not be horribly slow. Also, it should still
render properly and fill in all data.
On prod: http://ethicsdata.socrata.com/dataset/OMB-Appropriations-Earmarks-FY-2005-2008-2010/xvjq-vj6t

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>