NOOP: Consolidate before_filter checks by common check
The AdministrationController has a number of sections for managing
similar areas of content on the admin pages, where the access checks
within the section are the same. Rather than declare the necessary
before_filters on each method, I've grouped them by section.

QA: As with my prior commit, you'll want to re-test that all the
functionality on the /admin pages continue to work.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>