SECURITY: Lock down the AdministrationController properly.
Last night the GSA ran a scan of our environment using HP WebInspect,
which appeared to do a bunch of request fuzzing. Among the areas it
tested were URLs to the AdministrationController, and on inspection it
appears we've been Doing It Wrong.

The setup before had a few helper methods that we would call at the top
of each function to check various conditions were true: that a user has
a particular admin privilege, or a module is enabled, for example. These
helper functions would render an error if the condition wasn't met. The
problem is that it didn't stop the processing - at best, it'd continue
execution despite not having access to a given controller method, and in
the end you'd be presented with a DoubleRenderError. In some cases this
would be incorrect but secure, but in most cases this meant you could
bypass our Administration controller's security entirely if you knew
what you were doing.

This change converts the helper methods invocations to be used as
before_filter methods, and properly stops the invocation of the
controller method if access is denied. I tried to make this as
straightforward a port as possible for this change; another change after
this will clean up the before_filter invocations to not be scattered
throughout the code as much.

QA: Retest that the functionality in /admin works. There should be no
noticeable change, but pay particular attention to access controls - try
to do things with a user account that doesn't have a privileges, and
make sure you can't. Try to do things your account does have privileges
for, and ensure you still can.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>