FEATURE: Handle unpublished datasets * Don't show Save As, instead extra message on why it can't be saved * Extra text, icon & explanation next to title to clarify * Link back to published * Hide sharing, view creation elements ** Share links ** More Views ** Visualize creates (map, calendar, chart) ** Embed (form & SDP) ** Subscribe ** Discuss * Button to publish & redirect to published version * Changed 'Edit Dataset' in published to be all JS
QA:
After clicking 'Edit Dataset' on a dataset, you should be redirected
to a different URL that has mostly the same data, but has a big
[Working Copy] in front of the name and a different icon. This indicates
the unpublished version. This version cannot be shared or published, and
you can't create views on it. But you can edit it (both data and
columns), and then publish it, at which point all those changes will
appear on the published version (which has a consistent URL, and is the
one that appears in your profile & catalog).

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>