FEATURE: Rich catalog layout * New variant for browse that has a richer title/description/icon display * Use view type toggle to switch * Each column in browse can be toggled individually * Admin option to set default catalog view type * Refactored styles & JS for simple radio buttons in admin * Added preferred_image method to view.rb to get a nice icon/image for a view
QA:
In browse you should get view type toggles next to the sort dropdown.
This allows you to view the catalog in an expanded list format that
shows the full description and an icon for a dataset. If a dataset has a
thumbnail, it will use that; otherwise it will use the datatype icon (if
it has an iconUrl set it, that will be used first; but I'm not sure any
such datasets exist on test yet). In the Homepage admin section, there
is also an option to toggle the default catalog display.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>