FEATURE: Add cell-level commenting * When selecting a cell, button for comments appears as part of selection * Opens sidebar specific to comments for that cell * Load comments by row, then store them by row & column in Dataset * Highlight cell that currently has comments shown * feed-list is a bit more self-contained; mainly fixed up events * Base sidebar has option for hide callback on a pane * Table has support for adding arbitrary classes to a cell * No results message in feed sidebar * Must have cell_comments module enabled
QA:
Add feature cell_comments to the domain you want to test on.
When you select (not edit) a cell in the grid, there will be an extra
button to show comments for that cell. It will pop up a sidebar that
shows only comments for that cell. Test adding a comment; make sure it
only appears for that cell. The cell that corresponds to the comments
shown will be indicated in orange. Make sure comments appear properly on
reload.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>