FEATURE: Make the frontend respect the core server session timeouts
Up until now, we'd hard-coded the expiration time for the core server
session cookie (6 hours) in the frontend. We're about to introduce a
core server feature that lets some accounts have a different expiration
time, and we need the frontend to honor/respect it, so this change does
that.

QA: You can't really test this independently of the core server change,
which enables you to have some accounts with a much shorter session
timeout. What I'm looking for on the frontend side is if things simply
don't work at all - a "Something's Wrong" message while logging in our
out, for example.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>