BUG: Fix spots in the app that were giving mixed-content warnings
QA: This is hard to test because you'll never see what it was like
before this. The key thing this commit does is make it so that we're not
requesting HTTP content from an HTTPS page. Different browsers have
different methods of handling this, but IE is the most "strict" about it
- if this happens, you'll see a dialog that says:

  Do you want to view only the webpage content that was delivered
  securely?

or on older versions of IE:

  This page contains both secure and nonsecure items. Do you want to
  display the nonsecure items?

In browsers like Firefox, you'll generally see a box to the left of the
URL with the hostname or company name with a blue background behind it.
If there's a mixed-content violation, the blue box would go away. If
this happens, then it's a bug and I didn't catch every spot. Let me know
the URL of the page(s) with this issue in order to track them down.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>