FEATURE: Fixes to enable ESRI maps to load over HTTPS
This changes a bunch of references to ESRI maps to use HTTPS instead of
HTTP. Notably, it changes the default basemaps available to use HTTPS,
which enables the tiles to be served from a secure connection.

*****************************
* IMPORTANT DEPLOYMENT NOTE *
*****************************

This change does NOT fix existing maps. When creating or updating a map
view, the basemap layers are actually saved as JSON in the database. At
the time of this writing, there are about 180 views in production
already using http:// basemaps that will need to be manually updated:

blist_prod=> select count(*) from lens_display_format where format_key = 'layers' and value like '%"http://server.arcgisonline.com/%';
 count
-------
   176

There are an additional 4 basemaps that are using customer's custom ESRI
servers that we'll need to verify are availble via HTTPS before we
actually move them over:

blist_prod=> select count(*) from lens_display_format where format_key =
'layers' and value not like '%"http://server.arcgisonline.com/%';
 count
-------
     4

QA: Creating NEW ESRI maps with the upcoming HTTPS support should now
not pop up mixed-content warnings in IE, or in a browser like recent
versions of Firefox you'll see the SSL certificate name immediately to
the left of the URL in the address bar.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>