NOOP: Change how caching is done to account for content_for.
QA: Test the homepage browse control (ONLY the homepage one,
not the one at /browse). Make sure that:
1. It still caches. So, for example if you leave
a comment on one of the top ten datasets on the
homepage datasets browse, the comment count shouldn't
increase for a few minutes.
2. It still works. Most specifically, make sure that
the styling and the javascript still work (to test
the javascript, just make sure that the arrow next to
the dataset still expands and contracts the details
as expected). Clear your browser cache between each
load to make sure you're getting a fresh copy.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>