BUG: 4256 Error accessing federated datasets
This is a bug in View#href - it simply wasn't being smart about figuring
out the protocol for a given View.

Note that this isn't really a full fix for federation. If everyone
switches to HTTP simultaneously (as it's currently setup), everything
will work fine. If we try to setup domains to use HTTPS on a
case-by-case basis and federate between HTTP and HTTPS domains, this
will fail because the code that passes in the the port uses request.port
rather than actually looking up the port for the federation.

QA: Retest

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>