FEATURE: Add Resubmit for routing & approval * Owner/update_view rightsholders can Resubmit a rejected dataset * Update language & icons to handle Resubmission * approve_view takes a type, not just yes/no * Refactored rendering actions for queue & About * Resubmit also uses comment box like Reject
QA:
In routing and approval, when a dataset is rejected, the owner should
now get the option to Resubmit (only if the owner doesn't have Approve
priveleges). Resubmission takes a reason, like Reject. It should show up
in the history, but should cancel out the red rejected dot in the icon,
and be back to ready for approval for this stage.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>