FEATURE: Per-domain HTTPS-everywhere support
I ended up needing to fork the rack-ssl-enforcer gem, because we need to
have some custom logic as to whether a domain requires SSL. This isn't
the type of thing that's useful upstream, so I killed off the Gem and
just put the source in lib/rack and modified to suit.

Compared to production (but not a recent staging candidate) this means
we can flag a particular domain as "HTTPS everywhere" which, when
enabled, does the following:
* All requests to the site on HTTP simply redirect to the HTTPS version.
* Session cookies on these domains are marked as secure (and therefore
  only sent over an HTTPS connection)
* Non-enabled domains MAY use SSL (in fact, some pages, like /login,
  still require it) but the cookie is not marked as secure and the
  redirects between HTTP and HTTPS aren't enforced. Right now there's
  very little effort made to ensure a visitor stays on HTTP or HTTPS as
  they browse.

Compared to the recent staging versions with HTTPS everywhere enabled
for all domains:
* For domains that are turned on, there should be no noticeable change.
* For domains that are not turned on (the default), it should act more
  like the old production setup.

QA: You'll want to do a general regression test, particularly around
navigation between pages, authentication, and federated datasets. After
this deploys, I'll need to make some changes internally to pick a few
SSL-enforced domains; remind me to do this if you haven't heard from me
around when the candidate comes out!

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>