BUG: HTTPS-everywhere was not setting session cookie as HTTPS.
The order of the rack middleware here became a problem - the SSL
Enforcer needs to run relatively low in the middleware chain, so it can
inspect all the cookies in the response and potentially rewrite them to
append "Secure" on each one. However, re-ordering them exposed a
chicken/egg problem:

1. The Rack::SslEnforcer must come before the BlistCookieStore so that it
can rewrite the cookie the BlistCookieStore creates.
2. The CurrentDomainMiddleware must come before the Rack::SslEnforcer
because we use the current domain to determine whether we need to
enforce HTTPS.
3. The BlistCookieStore needs to come bofer the CurrentDomainMiddleware,
because we inspect the rack session to optionally override the
custom_site_config to preview a non-active theme. This causes a
dependency loop!

The solution is to split the CurrentDomainMiddleware such that the
custom_site_config is split into its own middleware and runs later.
According to Aiden, this isn't really even used any more, so the next
commit will be me removing this logic - but I wanted to keep the commits
separate.

QA: If you inspect the cookies on an HTTPS-enforced domain, they should
have the Secure flag set. On a non-HTTPS-enforced domain, they will not
have the secure flag set. (You can see this detail is Firefox using the
Web Developer addin/toolbar.) This is related to bug 4366, so it's worth
doing a regression on that.

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>